% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio.R
\name{bio_}
\alias{bio}
\alias{bio_}
\title{Simple query from biographic tables}
\usage{
bio_(tbl, .dots, env = parent.frame())

bio(tbl, ..., env = parent.frame())
}
\arguments{
\item{tbl}{the name of the bio table, e.g. degrees or activities}

\item{env}{symbol evaluation environment}

\item{...}{filter conditions}
}
\description{
Simple query from biographic tables
}

