% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_query.R
\name{custom_query}
\alias{custom_query}
\title{Flexible template for queries with single-column results}
\usage{
custom_query(custom, id_field, id_type, where = NULL, having = NULL)
}
\arguments{
\item{custom}{A SQL query (as a character vector of length 1)}

\item{id_field}{the name of the id column in \code{custom}}

\item{id_type}{a text label for the type of id}

\item{where}{pre-aggregated filter conditions}

\item{having}{post-aggregated filter conditions}
}
\description{
This functions provides a shorthand for queries of the form
"select x from (standalone-query) where ... group by ... having ..."
}
\details{
This function is similar to \code{\link{aggregate_q}} except
the source is not constrained to being a single table, it can be any query.
}
