% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flist.R
\name{flist}
\alias{flist}
\alias{flist_}
\title{Convert a list of Ids to a new list of a possibly different type}
\usage{
flist(savedlist, table, from, to, id_type = "cads_id", ..., schema = "CDW",
  env = parent.frame())

flist_(savedlist, table, from, to, id_type = "cads_id", .dots,
  schema = "CDW", env = parent.frame())
}
\arguments{
\item{savedlist}{a listbuilder list}

\item{table}{the table within which the flist will happen}

\item{from}{the origin ID field}

\item{to}{the destination ID field}

\item{...}{any conditions, a set of boolean expressions}

\item{schema}{"CDW" by default, no need to change it}

\item{env}{enviornment for evaluating table, from, to, and ...}
}
\description{
A list of allocations can be converted to a list of entities, via the
question "who gave to allocations in this list?" A list of entities can be
converted to a new list of entities via the question "Who are the children of
these entities?" And so on. \code{flist} is the general-purpose tool to make
these conversions.
}

