% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{operate}
\alias{\%and\%}
\alias{\%minus\%}
\alias{\%or\%}
\alias{operate}
\title{Combine two lists via a set operator}
\usage{
operate(block1, block2, operator)

block1 \%and\% block2

block1 \%or\% block2

block1 \%minus\% block2
}
\arguments{
\item{block1}{a listbuilder list}

\item{block2}{a listbuilder list}

\item{operator}{a combination operator, one of "intersect," "union," or "minus"}
}
\description{
Combine two lists via a set operator
}

