% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people.R
\name{people_}
\alias{people}
\alias{people_}
\title{Basic entity query}
\usage{
people_(.dots, include_deceased = FALSE, env = parent.frame())

people(..., include_deceased = FALSE, env = parent.frame())
}
\arguments{
\item{include_deceased}{should deceased individuals be included? (Defaults to FALSE)}

\item{env}{evaluation environment}

\item{...}{any conditions}
}
\description{
Basic entity query
}

