% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cdw.R
\docType{import}
\name{reexports}
\alias{get_cdw}
\alias{reexports}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{getcdw}{\code{\link[getcdw]{get_cdw}}}
}}
\keyword{internal}

