% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sql.R
\name{to_sql}
\alias{to_sql}
\title{Translate a listbuilder object into SQL}
\usage{
to_sql(obj)
}
\arguments{
\item{obj}{the listbuilder object to be converted}
}
\description{
Translate a listbuilder object into SQL
}
