% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_q.R
\name{aggregate_q_}
\alias{aggregate_q_}
\title{Template for single-table aggregate queries}
\usage{
aggregate_q_(table, where, having, id_field = "entity_id",
  id_type = "entity_id", schema = "CDW", env = parent.frame())
}
\arguments{
\item{table}{full name of the CDW table (eg d_entity_mv)}

\item{where}{pre-aggregated filter conditions}

\item{having}{post-aggregated filter conditions}

\item{id_field}{the name of the id column in \code{table}}

\item{id_type}{a text label for the type of id}

\item{schema}{the schema}

\item{env}{evaluation environment}
}
\description{
This functions provides a shorthand for queries of the form
"select x from y where ... group by ... having ..."
}

