% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_q.R
\name{simple_q}
\alias{simple_q}
\alias{simple_q_}
\title{Template for simple single-table queries.}
\usage{
simple_q(table, ..., id_field = "entity_id", id_type = "cads_id",
  schema = "CDW", env = parent.frame())

simple_q_(table, where, id_field = "entity_id", id_type = "cads_id",
  schema = "CDW", env = parent.frame())
}
\arguments{
\item{table}{full name of the CDW table (eg d_entity_mv)}

\item{...}{filter conditions}

\item{id_field}{the name of the id column in \code{table}}

\item{id_type}{a text label for the type of id}

\item{schema}{the schema}

\item{env}{evaluation environment}
}
\description{
This functions provides a shorthand for queries of the form "select x from y where ..."
}

